AE-1780: Flight phases/tests migration and clean-up
