+++ analysis_engine/flight_phase.py
@@ -16,19 +16,15 @@ from analysis_engine.library import (
     bearing_and_distance,
     cycle_finder,
     find_low_alts,
-    filter_slices_duration,
     first_order_washout,
     first_valid_sample,
     heading_diff,
     index_at_value,
-    is_index_within_slices,
     is_index_within_slice,
     is_slice_within_slice,
     last_valid_sample,
     mask_outside_slices,
     max_value,
-    moving_average,
-    nearest_neighbour_mask_repair,
     peak_curvature,
     rate_of_change,
     rate_of_change_array,
@@ -36,11 +32,8 @@ from analysis_engine.library import (
     runs_of_ones,
     shift_slice,
     shift_slices,
-    slices_above,
     slices_and,
     slices_and_not,
-    slices_below,
-    slice_duration,
     slices_extend_duration,
     slices_from_to,
     slices_int,
@@ -54,7 +47,7 @@ from analysis_engine.library import (
 
 from analysis_engine.node import (
     A, App, FlightPhaseNode, P, S, KTI, KPV, M,
-    aeroplane, aeroplane_only, helicopter, helicopter_only)
+    aeroplane, aeroplane_only, helicopter)
 
 from analysis_engine.settings import (
     AIRBORNE_THRESHOLD_TIME,
@@ -80,20 +73,9 @@ from analysis_engine.settings import (
     VERTICAL_SPEED_FOR_CLIMB_PHASE,
     VERTICAL_SPEED_FOR_DESCENT_PHASE,
 
-    AIRBORNE_THRESHOLD_TIME_RW,
-    AUTOROTATION_SPLIT,
-    HOVER_GROUNDSPEED_LIMIT,
-    HOVER_HEIGHT_LIMIT,
-    HOVER_MIN_DURATION,
-    HOVER_MIN_HEIGHT,
-    HOVER_TAXI_HEIGHT,
     LANDING_COLLECTIVE_PERIOD,
     LANDING_HEIGHT,
     LANDING_TRACEBACK_PERIOD,
-    TAKEOFF_PERIOD,
-    ROTOR_TRANSITION_ALTITUDE,
-    ROTOR_TRANSITION_SPEED_LOW,
-    ROTOR_TRANSITION_SPEED_HIGH,
 )
 
 
@@ -156,32 +138,6 @@ class Airborne(FlightPhaseNode):
                     if (duration / alt_aal.hz) > AIRBORNE_THRESHOLD_TIME:
                         self.create_phase(shift_slice(slice(begin, end),
                                                           start_point))
-
-
-class Autorotation(FlightPhaseNode):
-    '''
-    Look for at least 1% difference between the highest power turbine speed
-    and the rotor speed.
-    This is bound to happen in a descent, and we define the autorotation
-    period as from the initial onset
-    to the final establishment of normal operation.
-
-    Note: For Autorotation KPV: Detect maximum Nr during the Autorotation phase.
-    '''
-
-    can_operate = helicopter_only
-
-    def derive(self, max_n2=P('Eng (*) N2 Max'),
-               nr=P('Nr'), descs=S('Descending')):
-        for desc in descs:
-            # Look for split in shaft speeds.
-            delta = nr.array[desc.slice] - max_n2.array[desc.slice]
-            split = np.ma.masked_less(delta, AUTOROTATION_SPLIT)
-            split_ends = np.ma.clump_unmasked(split)
-            if split_ends:
-                self.create_phase(shift_slice(slice(split_ends[0].start,
-                                                    split_ends[-1].stop ),
-                                              desc.slice.start))
 
 
 class GoAroundAndClimbout(FlightPhaseNode):
@@ -828,88 +784,6 @@ class GearRetracted(FlightPhaseNode):
         self.create_phases(slices_remove_small_bits(gear_up))
 
 
-class Hover(FlightPhaseNode):
-    @classmethod
-    def can_operate(cls, available, ac_type=A('Aircraft Type')):
-        return ac_type == helicopter and \
-               all_of(('Altitude AGL', 'Airborne', 'Groundspeed'), available)
-
-    def derive(self, alt_agl=P('Altitude AGL'),
-               airs=S('Airborne'),
-               gspd=P('Groundspeed'),
-               trans_hfs=S('Transition Hover To Flight'),
-               trans_fhs=S('Transition Flight To Hover')):
-
-        low_flights = []
-        hovers = []
-
-        for air in airs:
-            lows = slices_below(alt_agl.array[air.slice], HOVER_HEIGHT_LIMIT)[1]
-            for low in lows:
-                if np.ma.min(alt_agl.array[shift_slice(low, air.slice.start)]) <= HOVER_MIN_HEIGHT:
-                    low_flights.extend([shift_slice(low, air.slice.start)])
-
-        repaired_gspd = repair_mask(gspd.array, frequency=gspd.hz,
-                                    repair_duration=8, method='fill_start')
-
-        slows = slices_below(repaired_gspd, HOVER_GROUNDSPEED_LIMIT)[1]
-        low_flights = slices_and(low_flights, slows)
-        # Remove periods identified already as transitions.
-        for low_flight in low_flights:
-            if trans_fhs:
-                for trans_fh in trans_fhs:
-                    if slices_overlap(low_flight, trans_fh.slice):
-                        low_flight = slice(trans_fh.slice.stop, low_flight.stop)
-
-            if trans_hfs:
-                for trans_hf in trans_hfs:
-                    if slices_overlap(low_flight, trans_hf.slice):
-                        low_flight = slice(low_flight.start, trans_hf.slice.start)
-
-            hovers.extend([low_flight])
-
-        # Exclude transition periods and trivial periods of operation.
-        self.create_phases(filter_slices_duration(hovers, HOVER_MIN_DURATION, frequency=alt_agl.frequency))
-
-
-class HoverTaxi(FlightPhaseNode):
-    @classmethod
-    def can_operate(cls, available, ac_type=A('Aircraft Type')):
-        return ac_type == helicopter and \
-               all_of(('Altitude AGL', 'Airborne', 'Hover'), available)
-
-    def derive(self, alt_agl=P('Altitude AGL'),
-               airs=S('Airborne'),
-               hovers=S('Hover'),
-               trans_hfs=S('Transition Hover To Flight'),
-               trans_fhs=S('Transition Flight To Hover')):
-
-        low_flights = []
-        air_taxis = []
-        taxis = []
-
-        if airs:
-            for air in airs:
-                lows = slices_below(alt_agl.array[air.slice], HOVER_TAXI_HEIGHT)[1]
-                taxis = shift_slices(lows, air.slice.start)
-        # Remove periods identified already as transitions.
-        if taxis:
-            for taxi in slices_and_not(taxis, [h.slice for h in hovers]):
-                if trans_fhs:
-                    for trans_fh in trans_fhs:
-                        if slices_overlap(taxi, trans_fh.slice):
-                            taxi = slice(trans_fh.slice.stop, taxi.stop)
-
-                if trans_hfs:
-                    for trans_hf in trans_hfs:
-                        if slices_overlap(taxi, trans_hf.slice):
-                            taxi = slice(taxi.start, trans_hf.slice.start)
-
-                air_taxis.extend([taxi])
-
-        self.create_phases(air_taxis)
-
-
 def scan_ils(beam, ils_dots, height, scan_slice, frequency,
              hdg=None, hdg_ldg=None, duration=10):
     '''
@@ -1318,48 +1192,6 @@ class Grounded(FlightPhaseNode):
             self._derive_helicopter(air, airspeed)
         else:
             self._derive_aircraft(speed, hdf_duration, air)
-
-
-class OnDeck(FlightPhaseNode):
-    '''
-    Flight phase for helicopters that land on the deck of a moving vessel.
-
-    Testing for motion will separate moving vessels from stationary decks, chosen as a better
-    option than testing the location against Google Earth for land/sea.
-
-    Also, movement was not practical as helicopters taxi at similar speeds to a ship sailing!
-
-    Note that this qualifies Grounded which is still asserted when On Deck.
-    '''
-
-    can_operate = helicopter_only
-
-    def derive(self, gnds=S('Grounded'),
-               pitch=P('Pitch'), roll=P('Roll')):
-
-        decks = []
-        for gnd in gnds:
-            # The fourier transform for pitching motion...
-            p = pitch.array[gnd.slice]
-            if np.all(p.mask):
-                continue
-            n = float(len(p)) # Scaling the result to be independet of data length.
-            fft_p = np.abs(np.fft.rfft(p - moving_average(p))) / n
-
-            # similarly for roll
-            r = roll.array[gnd.slice]
-            if np.all(r.mask):
-                continue
-            fft_r = np.abs(np.fft.rfft(r - moving_average(r))) / n
-
-            # What was the maximum harmonic seen?
-            fft_max = np.ma.max(fft_p + fft_r)
-
-            # Values of less than 0.1 were on the ground, and 0.34 on deck for the one case seen to date.
-            if fft_max > 0.2:
-                decks.append(gnd.slice)
-        if decks:
-            self.create_sections(decks)
 
 
 class Taxiing(FlightPhaseNode):
@@ -1834,19 +1666,6 @@ class RejectedTakeoff(FlightPhaseNode):
                         )
 
 
-class RotorsTurning(FlightPhaseNode):
-    '''
-    Used to suppress nuisance warnings on the ground.
-
-    Note: Rotors Running is the Multistate parameter, while Rotors Turning is the flight phase.
-    '''
-
-    can_operate = helicopter_only
-
-    def derive(self, rotors=M('Rotors Running')):
-        self.create_sections(runs_of_ones(rotors.array == 'Running'))
-
-
 class Takeoff(FlightPhaseNode):
     """
     This flight phase starts as the aircraft turns onto the runway and ends
@@ -2226,12 +2045,12 @@ class TCASOperational(FlightPhaseNode):
     There are different validity flags with different aircraft, and we need to make sure
     that TCAS does not operate on the ground. This phase merges the alternative sources
     to avoid repetition elsewhere.
-    
-    TCAS determines the approximate altitude of each aircraft above the ground. 
-    If this difference is less than 900 feet, TCAS considers the reporting aircraft 
+
+    TCAS determines the approximate altitude of each aircraft above the ground.
+    If this difference is less than 900 feet, TCAS considers the reporting aircraft
     to be on the ground.
     """
-    
+
     name = 'TCAS Operational'
     frequency = 1.0
 
@@ -2252,7 +2071,7 @@ class TCASOperational(FlightPhaseNode):
         if not operating:
             # No point in looking further if the aircraft didn't fly.
             return
-        
+
         if tcas_cc:
             # Build a list of the valid sections of Combined Control data...
             good_slices = []
@@ -2335,7 +2154,7 @@ class TCASTrafficAdvisory(FlightPhaseNode):
     def can_operate(cls, available):
         return any_one_of(('TCAS TA', 'TCAS All Threat Traffic', 'TCAS Traffic Alert', 'TCAS TA (1)'), available) \
             and 'TCAS Operational' in available
-    
+
     def derive(self, tcas_ops=S('TCAS Operational'),
                tcas_ta1=M('TCAS TA'),
                tcas_ta2=M('TCAS All Threat Traffic'),
@@ -2356,7 +2175,7 @@ class TCASTrafficAdvisory(FlightPhaseNode):
             tas_local = array.any_of('TA', 'Alert', ignore_missing=True)
             ta_slices = shift_slices(runs_of_ones(tas_local), tcas_op.slice.start)
             ta_slices = slices_remove_small_slices(ta_slices,
-                                                   time_limit=4.0, 
+                                                   time_limit=4.0,
                                                    hz=tcas_ta.frequency)
             all_slices.extend(ta_slices)
 
@@ -2370,29 +2189,29 @@ class TCASTrafficAdvisory(FlightPhaseNode):
                         to_pop.append(n)
             for pop in to_pop[::-1]:
                 all_slices.pop(pop)
-        
+
         self.create_phases(all_slices)
 
 class TCASResolutionAdvisory(FlightPhaseNode):
     '''
-    This uses the Combined Control parameter only because the TCAS RA signals are only 
-    present on aircraft with Combined Control as well, and the TCAS RA signals include 
+    This uses the Combined Control parameter only because the TCAS RA signals are only
+    present on aircraft with Combined Control as well, and the TCAS RA signals include
     the Clear of Conflict period, making the duration of the phase inconsistent.
     '''
-    
+
     @classmethod
     def can_operate(cls, available):
         return all_of(('TCAS Combined Control', 'TCAS Operational'), available) or \
                all_of(('TCAS RA', 'TCAS Operational'), available)
-    
+
     name = 'TCAS Resolution Advisory'
 
-    def derive(self, tcas_cc=M('TCAS Combined Control'), 
+    def derive(self, tcas_cc=M('TCAS Combined Control'),
                tcas_ops=S('TCAS Operational'),
                tcas_ra=M('TCAS RA')):
 
         for tcas_op in tcas_ops:
-            # We can be sloppy about error conditions because these have been taken 
+            # We can be sloppy about error conditions because these have been taken
             # care of in the TCAS Operational definition.
             if tcas_cc:
                 ra_slices = runs_of_ones(tcas_cc.array[tcas_op.slice].any_of(
@@ -2402,88 +2221,25 @@ class TCASResolutionAdvisory(FlightPhaseNode):
                     'Drop Track',
                     ignore_missing=True,
                 ))
-    
+
                 ra_slices = shift_slices(ra_slices, tcas_op.slice.start)
                 hz = tcas_cc.frequency
-                
+
             else:
                 # Operating with only a single TCAS RA signal, as recorded on some aircraft.
                 ra_slices = runs_of_ones(tcas_ra.array[tcas_op.slice].any_of(
                     'RA',
                     ignore_missing=True,
-                ))                
+                ))
                 ra_slices = shift_slices(ra_slices, tcas_op.slice.start)
                 hz = tcas_ra.frequency
-            
+
             # Where data is corrupted, single samples are a common source of error
             # time_limit rejects single samples, but 4+ sample events are retained.
             ra_slices = slices_remove_small_slices(ra_slices,
-                                                   time_limit=4.0, 
+                                                   time_limit=4.0,
                                                    hz=hz)
             self.create_phases(ra_slices)
-            
-                        
-################################################################################
-
-
-class TransitionHoverToFlight(FlightPhaseNode):
-    '''
-    The pilot normally makes a clear nose down pitching motion to initiate the
-    transition from the hover, and with airspeed built, will raise the nose and
-    initiate a clear climb to mark the end of the transition phase and start of the climb.
-    '''
-
-    can_operate = helicopter_only
-
-    def derive(self, alt_agl=P('Altitude AGL'),
-               ias=P('Airspeed'),
-               airs=S('Airborne'),
-               pitch_rate=P('Pitch Rate')):
-        for air in airs:
-            lows = np.ma.clump_unmasked(np.ma.masked_greater(alt_agl.array[air.slice],
-                                                             ROTOR_TRANSITION_ALTITUDE))
-            for low in lows:
-                trans_slices = slices_from_to(ias.array[air.slice][low],
-                                              ROTOR_TRANSITION_SPEED_LOW,
-                                              ROTOR_TRANSITION_SPEED_HIGH,
-                                              threshold=1.0)[1]
-                if trans_slices:
-                    for trans in trans_slices:
-                        base = air.slice.start + low.start
-                        ext_start = int(base  + trans.start - 20*ias.frequency)
-                        if alt_agl.array[ext_start]==0.0:
-                            trans_start = index_at_value(ias.array, 0.0,
-                                                         _slice=slice(base+trans.start, ext_start, -1),
-                                                         endpoint='first_closing')
-                        else:
-                            trans_start = np.ma.argmin(pitch_rate.array[ext_start:base+trans.start]) + ext_start
-                        self.create_phase(slice(trans_start, trans.stop+base))
-
-
-class TransitionFlightToHover(FlightPhaseNode):
-    '''
-    Forward flight to hover transitions are weakly defined from a flight parameter
-    perspective, so we only reply upon airspeed changes.
-    '''
-
-    can_operate = helicopter_only
-
-    def derive(self, alt_agl=P('Altitude AGL'),
-               ias=P('Airspeed'),
-               airs=S('Airborne'),
-               pitch_rate=P('Pitch Rate')):
-        for air in airs:
-            trans_slices = slices_from_to(ias.array[air.slice],
-                                          ROTOR_TRANSITION_SPEED_HIGH,
-                                          ROTOR_TRANSITION_SPEED_LOW,
-                                          threshold=1.0)[1]
-
-            if trans_slices:
-                for trans in shift_slices(trans_slices, air.slice.start):
-                    trans_end = index_at_value(ias.array, 0.0,
-                                                 _slice=slice(trans.stop, trans.stop+20*ias.frequency),
-                                                 endpoint='first_closing')
-                    self.create_phase(slice(trans.start, trans_end+1))
 
 
 class TurningInAir(FlightPhaseNode):
@@ -2553,18 +2309,18 @@ class ShuttlingApproach(FlightPhaseNode):
     '''
     Flight phase for the shuttling approach
     '''
-    
+
     def derive(self, approaches=App('Approach Information')):
         for approach in approaches:
             if approach.type == 'SHUTTLING':
                 self.create_section(approach.slice, name='Shuttling Approach')
-               
-                
+
+
 class AirborneRadarApproach(FlightPhaseNode):
     '''
     Flight phase for airborne radar approaches (ARDA/AROA)
     '''
-    
+
     def derive(self, approaches=App('Approach Information')):
         for approach in approaches:
             if approach.type == 'AIRBORNE_RADAR':
